remove approval from type mutation
